requirements.txt
# ============================================================================
# REQUIREMENTS - PREDICCIÓN DE CHURN EN TELECOMUNICACIONES
# ============================================================================

# Manipulación de datos
pandas>=1.5.0
numpy>=1.24.0

# Visualización
matplotlib>=3.6.0
seaborn>=0.12.0
plotly>=5.17.0

# Machine Learning
scikit-learn>=1.3.0
xgboost>=2.0.0
lightgbm>=4.1.0
catboost>=1.2.0

# Interpretabilidad
shap>=0.42.0

# Métricas y evaluación
imbalanced-learn>=0.11.0

# Jupyter y desarrollo
jupyter>=1.0.0
ipywidgets>=8.0.0

# Utilidades
tqdm>=4.65.0
joblib>=1.3.0
pickle5>=0.0.11; python_version < "3.8"

# API y deployment (opcional)
flask>=2.3.0
fastapi>=0.103.0
uvicorn>=0.23.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0

# Configuración
python-dotenv>=1.0.0
pyyaml>=6.0

# Logging
loguru>=0.7.0

# Streamlit y dependencias web
streamlit>=1.29.0